requirements.txt
apache-airflow==2.5.1
apache-airflow-providers-postgres==5.4.0
psycopg2-binary==2.9.5
requests==2.28.2
pandas==1.5.3
yfinance==0.2.12
python-dotenv==1.0.0